Updated user dashboard & settings
- Updated user dashboard
- Updated settings page
- Removed extra files
